Improve response-type parsing on grade report page
Refactor the way html fixes are applied on the grade report page.
Instead of attempting to generalize all fixes, we detect the type of
each response and apply specific fixes for that type.

This may cause bugs in response types we aren't aware of but should be
better in the long term.

Refs Bugzid #65225.